+++ src/ingestion_pipeline/chunking/document_chunker_factory.py
@@ -3,11 +3,11 @@
 import logging
 
 from ingestion_pipeline.chunking.base_document_chunker import DocumentChunker
-from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout_handler.types.key_value.layout_key_value import KeyValueChunker
-from ingestion_pipeline.chunking.layout_handler.types.layout_text import LayoutTextChunkingStrategy
-from ingestion_pipeline.chunking.layout_handler.types.list.list_chunker import LayoutListChunkingStrategy
-from ingestion_pipeline.chunking.layout_handler.types.table.layout_table import LayoutTableChunkingStrategy
+from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout.types.key_value.layout_key_value import KeyValueChunker
+from ingestion_pipeline.chunking.layout.types.layout_text import LayoutTextChunkingStrategy
+from ingestion_pipeline.chunking.layout.types.list.list_chunker import LayoutListChunkingStrategy
+from ingestion_pipeline.chunking.layout.types.table.layout_table import LayoutTableChunkingStrategy
 from ingestion_pipeline.chunking.line_sentence_chunker.config import LineSentenceChunkingConfig
 from ingestion_pipeline.chunking.line_sentence_splitter import LineBasedDocumentChunker
 from ingestion_pipeline.chunking.textract_layout_chunker import TextractLayoutDocumentChunker
