+++ src/ingestion_pipeline/chunking/layout/types/key_value/layout_key_value.py
@@ -8,7 +8,7 @@ from textractor.entities.key_value import KeyValue
 from textractor.entities.layout import Layout
 from textractor.entities.line import Line
 
-from ingestion_pipeline.chunking.layout_handler.types.base import ChunkingStrategyHandler
+from ingestion_pipeline.chunking.layout.types.base import ChunkingStrategyHandler
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
 from ingestion_pipeline.chunking.utils.bbox_utils import combine_bounding_boxes
