+++ src/ingestion_pipeline/chunking/layout/types/list/list_chunker.py
@@ -3,8 +3,8 @@
 import logging
 from typing import List, Optional
 
-from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout_handler.types.base import ChunkingStrategyHandler
+from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout.types.base import ChunkingStrategyHandler
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
 from ingestion_pipeline.chunking.utils.bbox_utils import combine_bounding_boxes
