+++ src/ingestion_pipeline/chunking/layout/types/table/layout_table.py
@@ -7,9 +7,9 @@ from textractor.entities.layout import Layout, Line
 from textractor.entities.table import Table
 
 from ingestion_pipeline.chunking.exceptions import ChunkException
-from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout_handler.types.base import ChunkingStrategyHandler
-from ingestion_pipeline.chunking.layout_handler.types.table.base import BaseTableChunker
+from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout.types.base import ChunkingStrategyHandler
+from ingestion_pipeline.chunking.layout.types.table.base import BaseTableChunker
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
 
 from .cell_chunker import CellTableChunker
