+++ src/ingestion_pipeline/chunking/layout/types/table/schemas.py
