+++ tests/chunking/strategies/list/test_list_chunker.py
@@ -9,8 +9,8 @@ import pytest
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../../..")))
 
-from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout_handler.types.list.list_chunker import LayoutListChunkingStrategy
+from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout.types.list.list_chunker import LayoutListChunkingStrategy
 from ingestion_pipeline.chunking.schemas import DocumentMetadata
