+++ tests/chunking/strategies/table/test_layout_table.py
@@ -7,10 +7,10 @@ from textractor.entities.layout import Layout, Line
 from textractor.entities.table import Table
 
 from ingestion_pipeline.chunking.exceptions import ChunkException
-from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout_handler.types.table import LayoutTableChunkingStrategy
-from ingestion_pipeline.chunking.layout_handler.types.table.cell_chunker import CellTableChunker
-from ingestion_pipeline.chunking.layout_handler.types.table.line_chunker import LineTableChunker
+from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout.types.table import LayoutTableChunkingStrategy
+from ingestion_pipeline.chunking.layout.types.table.cell_chunker import CellTableChunker
+from ingestion_pipeline.chunking.layout.types.table.line_chunker import LineTableChunker
 
 
 @pytest.fixture
